First try at reorganizing the code
