+++ popup.css
@@ -1,6 +1,6 @@
 :root {
     --input_scale_x: 1;
-    --input_scale_y: 0.8;
+    --input_scale_y: 0.7;
 }
 
 html * {
