+++ popup.html
@@ -1,3 +1,4 @@
+<!DOCTYPE html>
 <html>
     <head>
         <link rel="stylesheet" href="popup.css">
@@ -144,6 +145,8 @@
                 <div id="footer" class="clickable">Join us on Discord for free credits!</div>
             </div>
         </div>
+        <script src="utils.js"></script>
+        <script src="content_scripts.js"></script>
         <script src="popup.js"></script>
     </body>
 </html>
