+++ version/chrome/manifest.json
@@ -3,6 +3,9 @@
     "permissions": [
         "declarativeNetRequest"
     ],
-    "background": {"service_worker": "background.js"},
+    "background": {
+        "service_worker": "background.js",
+        "type": "module"
+    },
     "host_permissions": ["<all_urls>"]
 }
