+++ version/firefox/manifest.json
@@ -6,6 +6,11 @@
         "<all_urls>"
     ],
     "background": {
-        "scripts": ["background.js"]
+        "page": "page.html"
+    },
+    "browser_specific_settings": {
+        "gecko": {
+            "id": "{2f67aecb-5dac-4f76-9378-0ac4f2bedc9c}"
+        }
     }
 }
