+++ version/firefox/page.html
@@ -1,0 +1,8 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<script src="background.js" type="module"></script>
+</head>
+<body>
+</body>
+</html>
